Fußball

+++ Fußball/src/schritt1/Trainer.java
@@ -5,9 +5,9 @@ import schritt3.Person;
 public class Trainer extends Person {
 
 
-    private double erfahrung;
+    private int erfahrung;
 
-    public Trainer(String name, int alter, double erfahrung){
+    public Trainer(String name, int alter, int erfahrung){
         super(name,alter);
         this.erfahrung = erfahrung;
     }
@@ -18,7 +18,7 @@ public class Trainer extends Person {
         return erfahrung;
     }
 
-    public void setErfahrung(double erfahrung) {
+    public void setErfahrung(int erfahrung) {
         this.erfahrung = erfahrung;
     }
